requirements.txt
django
flake8
Pillow
pyexcel-ods
requests
# drf and recommend dependencies
djangorestframework
coreapi
Markdown
Pygments
django-filter
django-guardian
# django-tastypie
python-dateutil
python-mimeparse
django-tastypie